dw-us-bank-custody-doc-us-es4-function/requirements.txt
functions-framework==3.*
pandas
numpy
pysftp
openpyxl
paramiko
pyarrow
fastparquet
gcsfs
fsspec
openpyxl
lxml
requests
google-cloud-secret-manager
google-cloud-storage
google-cloud-logging